Disable OpenSearch autotune by default
This feature is only supported by some set of more expensive tiers, so
if someone wants to use a tier that doesn't support auto-tune, they'll
have to include additional configuration to the resource to support that
instead of the other way around.

By not enabling that feature by default, we'll only have it if someone
intentionally wants to use it.